Uploading the code
The application serves as a simple version of social media
network for multiple users. This application allows users to login and logout, and also create
a profile for themselves with desired profile picture and information about them. In the
application users can also create posts containing images with a text caption. Users can follow
and unfollow each other. There is a timeline that shows the 50 posts of the people the user is
following and also his posts. Users can search for each other using their names and see their
profile page.